pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hvac-air-quality"
version = "0.2.0-dev"
description = "Track HVAC filter efficiency to prevent asthma triggers"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Ming Yang"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Home Automation",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
keywords = ["air-quality", "hvac", "monitoring", "health", "iot"]
dependencies = [
    "requests>=2.32.3",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "plotly>=5.24.1",
    "jupyter>=1.1.1",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.1",
    "gspread>=6.1.4",
    "google-auth>=2.36.0",
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "schedule>=1.2.2",
]

[project.scripts]
# Scripts remain in root for Unifi Gateway compatibility
# Uncomment when files are moved to package structure
# hvac-collect = "hvac_air_quality.collect:main"
# hvac-analyze = "hvac_air_quality.analyze:main"
# hvac-read-sheets = "hvac_air_quality.sheets_reader:main"

[project.urls]
Homepage = "https://github.com/minghsuy/hvac-air-quality-analysis"
Documentation = "https://github.com/minghsuy/hvac-air-quality-analysis/wiki"
Issues = "https://github.com/minghsuy/hvac-air-quality-analysis/issues"
Changelog = "https://github.com/minghsuy/hvac-air-quality-analysis/blob/main/CHANGELOG.md"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "black>=24.10.0",
    "ruff>=0.7.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]

# Build configuration for when package structure is implemented
# [tool.hatch.build]
# sources = ["src"]
# 
# [tool.hatch.build.targets.wheel]
# packages = ["src/hvac_air_quality"]
